refactor(frontend): organize the folder structure ml-301
