+++ apps/frontend/src/libs/components/meeting-pdf/libs/helpers/meeting-pdf.styles.ts
