+++ apps/frontend/src/libs/components/meeting-pdf/meeting-pdf.tsx
@@ -4,8 +4,8 @@ import { Document, Page, Text, View } from "~/libs/components/components.js";
 import { formatDate } from "~/libs/helpers/helpers.js";
 import { type MeetingPdfProperties } from "~/libs/types/types.js";
 
-import { styles } from "./libs/meeting-pdf.styles.js";
-import { renderMarkdown } from "./libs/render-markdown.js";
+import { renderMarkdown } from "./libs/components/render-markdown.js";
+import { styles } from "./libs/helpers/meeting-pdf.styles.js";
 
 const MeetingPdf: React.FC<MeetingPdfProperties> = ({
 	actionItems,
